Add waiters to create task
Previously, calling createTask without any waiters would cause the
integration tests to fail since we depend on eventual consistency to
make assertions on the state of the service. There were issues with the
DescribeService calls being made hitting different DB replicas, causing
the integration tests to be brittle. This keeps the change that combines
the API calls on `compose service up` for a new service, while also
ensuring that the service gets to a stable state before further action.

This also adds an identifier prefix on the name of the service being
launched by the TestCreateClusterWithFargateService integ test.